+++ src/lib/validate-result.mjs
@@ -1,4 +1,4 @@
-import validateDecl from './validate-decl';
+import validateDecl from './validate-decl.mjs';
 
 // validate the css root
 export default (result, customProperties) => {
@@ -11,7 +11,7 @@ export default (result, customProperties) => {
 };
 
 // match custom property inclusions
-const customPropertyReferenceRegExp = /(^|[^\w-])var\([\W\w]+\)/;
+const customPropertyReferenceRegExp = /(^|[^\w-])var\([\W\w]+\)/i;
 
 // whether a declaration references a custom property
 const hasCustomPropertyReference = decl => customPropertyReferenceRegExp.test(decl.value);