+++ test/dummy-module-package/import-custom-properties.js
@@ -1,0 +1,5 @@
+export default {
+	customProperties: {
+		'--brand-white': 'white',
+	},
+};